polarbot-nmt-rest/requirements.txt
futures
gunicorn
flask
unidecode
http://download.pytorch.org/whl/cu75/torch-0.2.0.post1-cp35-cp35m-manylinux1_x86_64.whl
torchvision
nltk
sacremoses
